+++ assets/_Game/Script/Component/Object/ItemTypeEnum.ts
@@ -1,0 +1,19 @@
+import { _decorator, Component, Node } from 'cc';
+const { ccclass, property } = _decorator;
+
+export enum ItemType
+{
+    //FoodDrinks
+    Hamberger = 'Hamberger',
+    IceCreamCubBlue = 'IceCreamCubBlue',
+    Donut_Pink = 'Donut_Pink', 
+}
+
+
+
+@ccclass( 'ItemTypeEnum' )
+export class ItemTypeEnum
+{
+}
+
+