+++ assets/_Game/Script/Factory/ObjectType.ts
@@ -1,7 +1,6 @@
 export enum ObjectType {
-    CUBE = 'Cube',
-    SPHERE = 'Sphere',
-    APPLE = 'Apple',
-    GVS = 'GVS',
     // Thêm các object types khác ở đây khi cần
+    Hamberger = 'Hamberger',
+    IceCreamCubBlue = 'IceCreamCubBlue',
+    Donut_Pink = 'Donut_Pink', 
 }